pyproject.toml
[tool.poetry]
name = "imprint"
version = "0.1.0"
description = "Nostr long-form publishing app"
authors = ["AI Developer <dev@example.com>"]
license = "MIT"
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.30.0"}
jinja2 = "^3.1.4"
sqlalchemy = {version = "^2.0.30", extras = ["asyncio"]}
aiosqlite = "^0.20.0"
markdown2 = "^2.4.12"
websockets = "^12.0"
pydantic = "^2.7.0"
python-dotenv = "^1.0.1"
bcrypt = "^4.1.3"
bidict = "^0.23.1"
bech32 = "^1.2.0"
ecdsa = "^0.18.0"
httpx = "^0.27.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.7"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
